requirements.txt
attrs>=19.3.0
amqp>=5.2.0
channels>=4.1.0
channels_redis==4.2.0
daphne>=4.1.2
decorator==5.1.1
Django>=5.0.6
django-celery-results==2.5.1
django-crum==0.7.9
django-jsonbfield==0.1.0
django-pglocks==1.0.4
django-polymorphic==3.1.0
django-qsstats-magic==1.1.0
django-rest-swagger==2.2.0
djangorestframework==3.15.1
django-rest-passwordreset==1.4.1
drf_yasg==1.21.7
kombu==5.3.7
celery==5.4.0
redis==5.0.5
paramiko>=3.4.0
pexpect==4.9.0
psutil>=5.9.8
psycopg2>=2.7.4
psycopg2-binary>=2.9.9
python-daemon==3.0.1
python-dateutil==2.9.0post0
prometheus-client
pymongo~=4.7.3
pytz==2024.1
PyYAML>=6.0.1
PyJWT>=2.8.0
requests>=2.32.3
simplejson==3.19.2
uritemplate==4.1.1
urllib3>=2.2.1
uWSGI==2.0.26
sphinxcontrib-httpdomain
setuptools_scm
jsonschema==4.22.0
pyparsing==3.1.2
pygments==2.18.0
ansiconv==1.0.0
tzcron==1.0.0
django-debug-toolbar==4.4.2
django-debug-toolbar-force==0.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
prometheus_client~=0.20.0
coreapi~=2.3.3
six~=1.16.0
Jinja2~=3.1.4
cryptography~=42.0.8
packaging~=24.0
asgiref~=3.8.1
setuptools~=70.0.0
django-auditlog